Replace RATA-CMIX entirely with fx4-cmix's release/google-cloud-hutter
Complete overwrite, at the user's explicit direction: the older RATA-CMIX
code (cmix-lex + fxcm_v26 + R1 reordering + ESN/NLMS correction, the
587,138,826-byte post-R1 stream) is superseded, not merged with. That
lineage is exactly what fx4-cmix's own project history already measured as
worse -- a full v26 adoption cost +15,409 bytes against a v22 baseline on
the 1% harness, which is why fx4-cmix's current stack goes v22-based and
additive instead (v22++) rather than v26-wholesale.

Mechanical: every tracked file removed, then git archive release/google-
cloud-hutter (from the local naveenbijalwan/fx4-cmix clone, one commit
ahead of origin -- a trivial .dict-tracking cleanup, strictly an
improvement) extracted in its place. Deletions were exactly the four
old-lineage-only files: esn-nlms.{cpp,h}, r1_reorder_transform.{cpp,h}.
Git detected genuine renames for the transformer sources (src/third_party/
fx2_transformer/ -> cpp_infer/src/), confirming shared underlying content
despite the reorganization.

Verified before this commit: make target93 builds clean and links to a
cmix binary byte-identical in size (414,648 bytes) to fx4-cmix's own
verified build.

+++ src/mixer/sigmoid.cpp
@@ -17,7 +17,7 @@ float Sigmoid::Logit(float p) const {
 }
 
 float Sigmoid::Logistic(float p) {
-  return 1.0f / (1.0f + expf(-p));
+  return 1 / (1 + exp(-p));
 }
 
 float Sigmoid::FastLogistic(float p) {
